Support ofborg comments inline with other text
The old design of the parser treated all whitespace the same and
mandated that `grahamcofborg` (plus the `@`) be the first token in the
text. This allowed for some ridiculous but command calls:

    grahamcofborg build foo
       bar
          baz

This used to become a build instruction for foo, bar, and baz. After
this change, it is just an instruction for building foo. This allows
for comments for people to be intertwined with comments for the bot:

    grahamcofborg build foo

    Let's see what happens!

Before this would unintentionally become a build instruction for
`foo`, `Let's`, `see`, `what`, `happens!`, and is now only going to
build `foo`.

Additionally, this comment would do nothing:

    Let's see what happens!
    grahamcofborg build foo

Or a more real case where people expected this to work:

    /cc grahamc for ^^
    GrahamcOfBorg eval

This will continue to not produce a build instruction, because
grahamcofborg must be the first word of a line:

    foo bar grahamcofborg build foo

Note: I've removed `@`s from all usernames to avoid accidental email.