Tag added/removed packages on PRs
If packages (by name attribute and architecture) are present in the
first eval and absent in the second eval, mark it as having cleanup.

If packages (by name attribute and architecture) are absent in the
first eval and present in the second eval, mark it as having a new
package.

Note, though, that the generic "has cleanup" tag is too generic to
automatically remove it if ofborg thinks it isn't true.

Closes #63