add new eval checker for the package list with allowAliases = false
closes #248

By checking the package list without aliases, we can make it more
likely that people do not introduce new aliases into `nixpkgs` without
removing them from the repository. Aliases are supposed to be for
backward compatibility.

Also, we can make sure this configuration option continues to work for
people. Noteably, this configuration option is used by
r-ryantm/nixpkgs-update to avoid submitting duplicate pull requests
for the package and its alias, and to use the correct (non-alaised)
name in the pull request.