tasks/eval/nixpkgs: pass in rev, use proper rev and shortrev
Here, we can easily access the real commit ID, so pass it in instead of
a bogus rev and shortRev.

Hydra seems to have passed in rev all the time, and nixpkgs will soon
require `rev` to be set.

This initially broke nixpkgs eval in
https://github.com/NixOS/nixpkgs/commit/5e8545e72341887bb371407a71a723bc0e9c7844,
so we now update ofborg to always pass `rev` when instantiating nixpkgs
to avoid breaking it.