WIP: Introduce an one-off-build command to reproduce ofBorg locally
Motivated by trying to review a change in how ofBorg reports build
errors I started hacking on a command that allowed me to execute just
the build part of ofBorg without setting up a local queue or config
files.

This is still a WIP that is hardcoded to exactly one PR and commit
revision. I plan to convert all the constants to Clap CLI arguments once
I moved the build system of the project to naersk and can then introduce
new dependencies.

Things to be done:

 - [ ] Move constants to Clap
   - [ ] The system attribute should be defaulted to the current system.
         We can use the `target_os` and `target_arch` values for that.
	 The user should still be able to override thos tho...
 - [ ] Make the logging a bit prettier
 - [ ] Implement a custom Notification receiver that just logs the
       messages as they come in.
 - [ ] (idea) add support for the eval part as well so we can run ofBorg
        locally just like we can run nixpkgs-review.