clone: attempt to clean git if checking out fails
Somehow, some evaluators have had dirty git checkouts, causing them to
continually crash. We can fix this by attempting to cleanup the checkout
if we fail to switch to a different commit. If checking out another
commit still fails after that, then it is acceptable to crash (as it is
indicative of another problem).